Rework pricing
